finish updating tree

+++ include/thundergbm/util/device_lambda.cuh
@@ -27,7 +27,7 @@ __global__ void lambda_2d_sparse_kernel(int len1, int *len2, L lambda) {
 
 ///p100 has 56 MPs, using 32*56 thread blocks
 template<unsigned int NUM_BLOCK = 32 * 56, unsigned int BLOCK_SIZE = 512, typename L>
-void device_loop(size_t len, L lambda) {
+void device_loop(int len, L lambda) {
     if (len > 0) {
         lambda_kernel << < NUM_BLOCK, BLOCK_SIZE >> > (len, lambda);
         CUDA_CHECK(cudaPeekAtLastError());
